+++ src/features/user/friends/index.ts
@@ -4,7 +4,6 @@ export { FriendsPanel } from "./components/FriendsPanel";
 // Hooks
 export { useFriendProfiles } from "./hooks/useFriendProfiles";
 export { useFriendshipStatus } from "./hooks/useFriendshipStatus";
-export { useUserFriendCount } from "./hooks/useUserFriendCount";
 export { useUserFriends } from "./hooks/useUserFriends";
 
 // Types